Refactoring category theory
