--- base/shared/src/main/scala/scalaz/ct/ApplicativeSyntax.scala
@@ -1,10 +1,0 @@
-package scalaz
-package ct
-
-import language.experimental.macros
-
-trait ApplicativeSyntax {
-  implicit final class ToApplicativeOps[A](a: A) {
-    def pure[F[_]](implicit ev: Applicative[F]): F[A] = macro meta.Ops.i_0
-  }
-}
