--- base/shared/src/main/scala/scalaz/ct/ApplyClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait ApplyClass[F[_]] extends FunctorClass[F] {
-  def ap[A, B](fa: F[A])(f: F[A => B]): F[B]
-}
