--- base/shared/src/main/scala/scalaz/ct/BifunctorInstances.scala
@@ -1,11 +1,0 @@
-package scalaz
-package ct
-
-trait BifunctorInstances {
-
-  implicit final val tuple2Bifunctor: Bifunctor[Tuple2] =
-    instanceOf(new BifunctorClass[Tuple2] with BifunctorClass.DeriveBimap[Tuple2] {
-      def lmap[A, B, S](fab: (A, B))(f: A => S): (S, B) = fab.copy(_1 = f(fab._1))
-      def rmap[A, B, T](fab: (A, B))(f: B => T): (A, T) = fab.copy(_2 = f(fab._2))
-    })
-}
