--- base/shared/src/main/scala/scalaz/ct/BindFunctions.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait BindFunctions {
-  def flatMap[M[_], A, B](ma: M[A])(f: A => M[B])(implicit M: Bind[M]): M[B] = M.flatMap(ma)(f)
-}
