--- base/shared/src/main/scala/scalaz/ct/BindInstances.scala
@@ -1,4 +1,0 @@
-package scalaz
-package ct
-
-trait BindInstances {}
