--- base/shared/src/main/scala/scalaz/ct/ComonadClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait ComonadClass[F[_]] extends CobindClass[F] {
-  def copoint[A](fa: F[A]): A
-}
