--- base/shared/src/main/scala/scalaz/ct/ComonadSyntax.scala
@@ -1,10 +1,0 @@
-package scalaz
-package ct
-
-import scala.language.experimental.macros
-
-trait ComonadSyntax {
-  implicit final class ToComonadOps[F[_], A](self: F[A]) {
-    def copoint(implicit ev: Comonad[F]): A = macro meta.Ops.i_0
-  }
-}
