--- base/shared/src/main/scala/scalaz/ct/ComposeClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package ct
-
-trait ComposeClass[=>:[_, _]] {
-  def compose[A, B, C](f: B =>: C, g: A =>: B): (A =>: C)
-}
