+++ base/shared/src/main/scala/scalaz/ct/applicative.scala
@@ -1,6 +1,8 @@
 package scalaz
 package ct
 
+import language.experimental.macros
+
 trait ApplicativeClass[F[_]] extends ApplyClass[F] {
   def pure[A](a: A): F[A]
 }
@@ -10,5 +12,10 @@ object ApplicativeClass {
   trait DeriveMap[F[_]] extends ApplicativeClass[F] with MonadClass.Alt[DeriveMap[F]] {
     final override def map[A, B](ma: F[A])(f: (A) => B): F[B] = ap(ma)(pure(f))
   }
+}
 
+trait ApplicativeSyntax {
+  implicit final class ToApplicativeOps[A](a: A) {
+    def pure[F[_]](implicit ev: Applicative[F]): F[A] = macro meta.Ops.i_0
+  }
 }
