+++ base/shared/src/main/scala/scalaz/ct/apply.scala
@@ -3,6 +3,10 @@ package ct
 
 import scala.language.experimental.macros
 
+trait ApplyClass[F[_]] extends FunctorClass[F] {
+  def ap[A, B](fa: F[A])(f: F[A => B]): F[B]
+}
+
 trait ApplySyntax {
   implicit final class ToApplyOps[F[_], A](fa: F[A]) {
     def ap[B](f: F[A => B])(implicit ev: Apply[F]): F[B] =
