+++ base/shared/src/main/scala/scalaz/ct/bifunctor.scala
@@ -1,6 +1,8 @@
 package scalaz
 package ct
 
+import language.experimental.macros
+
 /** A typeclass for types which are (covariant) [[Functor]]s in both type parameters.
  *
  * Minimal definition:
@@ -42,5 +44,23 @@ object BifunctorClass {
     val F: Bifunctor[F]
 
     def map[B, BB](ma: F[A, B])(f: B => BB): F[A, BB] = F.rmap(ma)(f)
+  }
+}
+
+trait BifunctorInstances {
+
+  implicit final val tuple2Bifunctor: Bifunctor[Tuple2] =
+    instanceOf(new BifunctorClass[Tuple2] with BifunctorClass.DeriveBimap[Tuple2] {
+      def lmap[A, B, S](fab: (A, B))(f: A => S): (S, B) = fab.copy(_1 = f(fab._1))
+      def rmap[A, B, T](fab: (A, B))(f: B => T): (A, T) = fab.copy(_2 = f(fab._2))
+    })
+}
+
+trait BifunctorSyntax {
+  implicit final class ToBifunctorOps[F[_, _]: Bifunctor, A, B](ma: F[A, B]) {
+    def bimap[S, T](f: A => S, g: B => T): F[S, T] = macro meta.Ops.f_2
+
+    def lmap[S](f: A => S): F[S, B] = macro meta.Ops.f_1
+    def rmap[T](f: B => T): F[A, T] = macro meta.Ops.f_1
   }
 }
