+++ base/shared/src/main/scala/scalaz/ct/category.scala
